refactor(mobile): extract FeatureIntroLayout to shared mvvm component
Move the feature intro sheet layout out of GenericAwarenessModal so it can
be reused by Backup Hub and other bottom sheets. Export resolveThemedImageUrl
from live-common for themed hero images.
