+++ apps/ledger-live-mobile/src/mvvm/components/FeatureIntroLayout/types.ts
@@ -1,0 +1,12 @@
+import type { GenericAwarenessModalFeatureIntro } from "@ledgerhq/live-common/genericAwarenessModal";
+
+export type FeatureIntroViewModel = Readonly<{
+  content: GenericAwarenessModalFeatureIntro;
+  onPrimaryPress: () => void;
+  onSecondaryPress: () => void;
+}>;
+
+export type FeatureIntroLayoutProps = Readonly<{
+  onClose: () => void;
+  viewModel: FeatureIntroViewModel;
+}>;
