+++ libs/ledger-live-common/src/genericAwarenessModal/resolveThemedImageUrl.ts
@@ -5,6 +5,12 @@ export type ThemedImageUrls = {
 
 export type ThemeVariant = "light" | "dark";
 
+/** Maps a single image URL to themed fields when the same asset is used in light and dark mode. */
+export const createThemedImageUrls = (imageUrl: string | undefined): ThemedImageUrls => ({
+  imageUrlLight: imageUrl ?? "",
+  imageUrlDark: imageUrl ?? "",
+});
+
 export const resolveThemedImageUrl = (urls: ThemedImageUrls, theme: ThemeVariant): string => {
   const imageUrlLight = urls.imageUrlLight.trim();
   const imageUrlDark = urls.imageUrlDark.trim();
